+++ event_data.cpp
@@ -1,4 +1,4 @@
 #include "event_data.h"
 
-EventData::EventData(const std::shared_ptr<SiteData> &site, IReactingRole *reaction, float rate) :
+EventData::EventData(const std::shared_ptr<SiteData> &site, IReactingRole *reaction, double rate) :
     _site(site), _reaction(reaction), _rate(rate) {}
