+++ treebasedsimulation_context.h
@@ -15,7 +15,7 @@ class TreeBasedSimulationContext : public SimulationBaseContext
 public:
     TreeBasedSimulationContext(AreaData *area);
 
-    float doReaction();
+    double doReaction();
 
 private:
     void updateRates(SiteNode *siteNode) const;
@@ -29,7 +29,7 @@ private:
 template <int treeBaseWidth>
 TreeBasedSimulationContext<treeBaseWidth>::TreeBasedSimulationContext(AreaData *area) : SimulationBaseContext(area) {
     throughArea([this](int *cell, int **neighbours) {
-        SiteNode *siteNode = new SiteNode(cell, neighbours); // was deleted into MCTree
+        SiteNode *siteNode = new SiteNode(cell, neighbours); // will be deleted into MCTree
         this->updateRates(siteNode);
         this->_cellsToNodes[cell] = siteNode;
         this->_tree.add(siteNode);
@@ -37,11 +37,11 @@ TreeBasedSimulationContext<treeBaseWidth>::TreeBasedSimulationContext(AreaData *
 }
 
 template <int treeBaseWidth>
-float TreeBasedSimulationContext<treeBaseWidth>::doReaction() {
-    float totalRate = _tree.sum();
+double TreeBasedSimulationContext<treeBaseWidth>::doReaction() {
+    double totalRate = _tree.sum();
     if (totalRate == 0) return 0;
 
-    float r = randomN01() * totalRate;
+    double r = randomN01() * totalRate;
     SiteNode *currentSiteNode = static_cast<SiteNode *>(_tree.find(&r)); // here r is decreasing when we finds necessary site
     int reactionIndex = currentSiteNode->reactionIndex(r);
